refactor: reorganize manage module imports and constants
